refactor(structure): reorganize project into proper package hierarchy
