+++ src/main/kotlin/racingcar/io/RaceDisplay.kt
@@ -1,8 +1,9 @@
-package racingcar
+package racingcar.io
+import racingcar.model.Car
 
 object RaceDisplay {
     fun displayLapProgress(cars: List<Car>, lapNumber: Int){
-        println("\nCars progress after round: $lapNumber")
+        println("\nCars progress after round $lapNumber:")
         cars.forEach { car ->
             println("${car.name} : ${"-".repeat(car.position)}")
         }
@@ -10,6 +11,6 @@ object RaceDisplay {
 
     fun displayRaceWinners(cars: List<Car>, leadPosition: Int){
         val winners = cars.filter { car -> car.position == leadPosition }.map { car -> car.name }
-        println("Winners: ${winners.joinToString(", ")}")
+        println("\nWinners: ${winners.joinToString(", ")}")
     }
 }
